+++ app/Article.php
@@ -11,6 +11,8 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\Blade;
+use Illuminate\Support\Str;
 
 class Article extends Model implements TranslatableContract
 {
@@ -140,5 +142,20 @@ class Article extends Model implements TranslatableContract
     public function getReadingTimeAttribute()
     {
         return getEstimatedReadingTime($this->summary . ' ' . $this->content, \App::getLocale());
+    }
+
+    public function getParsedContentAttribute()
+    {
+        $content = Str::of(html_entity_decode($this->content))->replaceMatches("/\[article_teaser id=(.*?)\]/", function ($match) {
+            if (empty($match[1])) return '';
+            $id = Str::of($match[1]);
+            $article = Article::published()->find($id);
+
+            if(!$article) return '';
+            
+            $article_teaser = Blade::render('<x-article_teaser :article="$article" />', ['article' => $article]);
+            return $article_teaser;
+        });
+        return $content;
     }
 }