+++ app/Importers/OglImporter.php
@@ -55,7 +55,7 @@ class OglImporter extends AbstractImporter
         return $id;
     }
 
-    protected function getItemImageFilenameFormat(array $record)
+    protected function getItemImageFilenameFormat(array $record): string
     {
         $filename = sprintf(
             '%s%s',
@@ -66,7 +66,7 @@ class OglImporter extends AbstractImporter
             $filename = sprintf('%s%s', $filename, $record['Lomeni_S']);
         }
 
-        return $filename . '{_*,}';
+        return sprintf('%s(_.*)?', preg_quote($filename));
     }
 
     protected function hydrateIdentifier(array $record)