+++ app/Importers/PnpKarasekImporter.php
@@ -2,10 +2,9 @@
 
 namespace App\Importers;
 
-use App\Import;
 use App\ImportRecord;
+use App\Item;
 use Illuminate\Support\Str;
-use SplFileInfo;
 
 class PnpKarasekImporter extends AbstractImporter
 {
@@ -47,16 +46,16 @@ class PnpKarasekImporter extends AbstractImporter
         };
     }
 
-    public function import(Import $import, SplFileInfo $file)
+    public function import(ImportRecord $import_record, $stream): array
     {
         $this->counter = 0;
-        return parent::import($import, $file);
+        return parent::import($import_record, $stream);
     }
 
-    protected function importSingle(array $record, Import $import, ImportRecord $import_record)
+    protected function importSingle(array $record, ImportRecord $import_record): Item
     {
         $this->counter++;
-        return parent::importSingle($record, $import, $import_record);
+        return parent::importSingle($record, $import_record);
     }
 
     protected function getItemId(array $record)
@@ -64,10 +63,10 @@ class PnpKarasekImporter extends AbstractImporter
         return sprintf('CZE:PNP.%s', $this->getSlug($record['Inventární číslo:']));
     }
 
-    protected function getItemImageFilenameFormat(array $record)
+    protected function getItemImageFilenameFormat(array $record): string
     {
         $slug = $this->getSlug($record['Inventární číslo:']);
-        return sprintf('%s{_*,}', $slug);
+        return sprintf('%s(_.*)?', preg_quote($slug));
     }
 
     protected function getSlug($identifier)