+++ public/ckeditor/config.js
@@ -16,6 +16,7 @@ CKEDITOR.editorConfig = function( config ) {
 		{ name: 'insert'},
 		{ name: 'forms' },
 		{ name: 'wuSlick' },
+		{ name: 'articleTeaser' },
 		{ name: 'tools' },
 		{ name: 'document',	   groups: [ 'mode', 'document', 'doctools' ] },
 		{ name: 'others' },
@@ -36,7 +37,7 @@ CKEDITOR.editorConfig = function( config ) {
 
 	// Simplify the dialog windows.
 	config.removeDialogTabs = 'image:advanced;link:advanced';
-	config.extraPlugins = 'slick';
+	config.extraPlugins = 'slick,articleTeaser';
 	config.allowedContent = true;
 	config.embed_provider = '//ckeditor.iframe.ly/api/oembed?url={url}&callback={callback}';