+++ public/ckeditor/plugins/articleTeaser/lang/sk.js
@@ -1,0 +1,8 @@
+'use strict'
+
+;(function (CKEDITOR) {
+    CKEDITOR.plugins.setLang('articleTeaser', 'sk', {
+        title: 'Ukážka',
+        idRequired: 'Zadajte id článku (číslo)',
+    })
+})(CKEDITOR)