+++ resources/views/frontend/articles/show.blade.php
@@ -108,7 +108,7 @@
             </div>
             <div class="row">
                 <div class="col-md-8 col-md-push-2 long-text">
-                    {!! $article->content !!}
+                    {!! $article->parsed_content !!}
                 </div>
             </div>
             @unless (Cookie::has('newsletterSubscribedAt'))