+++ tests/Importers/OglImporterTest.php
@@ -8,7 +8,6 @@ use App\Matchers\AuthorityMatcher;
 use App\Repositories\CsvRepository;
 use Illuminate\Contracts\Translation\Translator;
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use SplFileInfo;
 use Tests\TestCase;
 
 class OglImporterTest extends TestCase
@@ -194,7 +193,10 @@ class OglImporterTest extends TestCase
             $this->app->get(Translator::class)
         );
         $import = Import::create();
-        $items = $importer->import($import, new SplFileInfo(''));
+        $import_record = $import->records()->create([
+            'filename' => 'filename.csv',
+        ]);
+        $items = $importer->import($import_record, stream: null);
 
         $this->assertEquals('', $import->records()->first()->error_message);
         $this->assertCount(1, $items);
@@ -202,7 +204,7 @@ class OglImporterTest extends TestCase
         return $items[0];
     }
 
-    public function getData(array $overrides)
+    protected function getData(array $overrides)
     {
         return $overrides + [
             'Rada_S' => 'P',