+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+## [2.69.0] - 2022-11-08
+### Fixed
+- indexing of Authority roles
+
+### Changed
+- refactored import files
+
+### Added 
+- parse article teaser from admin
+
 ## [2.68.0] - 2022-10-13
 ### Added
 - default Scout config to enable Scout search functionality